15 clean up RecordGrid
